+++ src/MauiControlsExtras/Controls/Accordion/Accordion.xaml.cs
@@ -1,5 +1,6 @@
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
+using System.Diagnostics.CodeAnalysis;
 using System.Windows.Input;
 using MauiControlsExtras.Base;
 using MauiControlsExtras.Theming;
@@ -749,6 +750,7 @@ public partial class Accordion : HeaderedControlBase, IKeyboardNavigable, ISelec
     /// <summary>
     /// Initializes a new instance of the <see cref="Accordion"/> class.
     /// </summary>
+    [DynamicDependency(nameof(CurrentBorderColor), typeof(Accordion))]
     public Accordion()
     {
         InitializeComponent();